code-archives/requirements.txt
PyPDF2>=3.0.1
tqdm>=4.66.0
pillow>=10.0.0
requests>=2.31.0
